+++ pyproject.toml
@@ -122,8 +122,8 @@ omit = [
     "*/foundation/debuggable.py",
     "*/foundation/mapper.py",
     "*/foundation/ports.py",
-    "*/application/ports/outbound/notifier.py",
-    "*/application/ports/outbound/repository.py",
+    "*/application/ports/outbound/notifier_port.py",
+    "*/application/ports/outbound/repository_port.py",
     "**/inbound.py",
     "**/outbound.py",
     "**/specification_repository.py",
