+++ scripts/release/application/ports/outbound/release_command_bus.py
@@ -1,7 +1,7 @@
 from typing import Any, Generic, Protocol, TypeVar
 
 from forging_blocks.application.ports.inbound.message_handler import MessageHandler
-from forging_blocks.application.ports.outbound.message_bus import MessageBusPort
+from forging_blocks.application.ports.outbound.message_bus_port import MessageBusPort
 from forging_blocks.foundation.messages.command import Command
 
 CommandType = TypeVar("CommandType", bound=Command[Any], contravariant=True)
