+++ src/forging_blocks/application/ports/outbound/external_service_port.py
