+++ src/forging_blocks/application/ports/outbound/query_fetcher_port.py
