+++ src/forging_blocks/infrastructure/event_stores/in_memory_event_store.py
@@ -13,7 +13,7 @@ from uuid import UUID
 
 from forging_blocks.application.errors.concurrency_error import ConcurrencyError
 from forging_blocks.application.errors.event_store_error import EventStoreError
-from forging_blocks.application.ports.outbound.event_store import EventStorePort
+from forging_blocks.application.ports.outbound.event_store_port import EventStorePort
 from forging_blocks.foundation.messages.event import Event
 from forging_blocks.foundation.result import Err, Ok, Result
