+++ src/forging_blocks/infrastructure/file_system/os_file_system.py
@@ -7,7 +7,7 @@ and ``os`` calls, avoiding external dependencies like ``aiofiles``.
 import asyncio
 from pathlib import Path
 
-from forging_blocks.application.ports.outbound.file_system import FileSystemPort
+from forging_blocks.application.ports.outbound.file_system_port import FileSystemPort
 
 
 class OSFileSystem(FileSystemPort):
