+++ src/forging_blocks/infrastructure/message_bus/message_bus_command_sender.py
@@ -3,8 +3,8 @@
 Delegates command dispatch to an injected ``MessageBusPort``.
 """
 
-from forging_blocks.application.ports.outbound.command_sender import CommandSenderPort
-from forging_blocks.application.ports.outbound.message_bus import MessageBusPort
+from forging_blocks.application.ports.outbound.command_sender_port import CommandSenderPort
+from forging_blocks.application.ports.outbound.message_bus_port import MessageBusPort
 from forging_blocks.foundation.messages.command import Command
