+++ src/forging_blocks/infrastructure/message_bus/message_bus_event_publisher.py
@@ -3,8 +3,8 @@
 Delegates event publishing to an injected ``MessageBusPort``.
 """
 
-from forging_blocks.application.ports.outbound.event_publisher import EventPublisherPort
-from forging_blocks.application.ports.outbound.message_bus import MessageBusPort
+from forging_blocks.application.ports.outbound.event_publisher_port import EventPublisherPort
+from forging_blocks.application.ports.outbound.message_bus_port import MessageBusPort
 from forging_blocks.foundation.messages.event import Event
